chore: create app data_loading and move content from data_analysis to it
