+++ crested_myna/crested_myna/settings.py
@@ -61,6 +61,7 @@ BASE_APPS = [
 
 LOCAL_APPS = [
     'data_analysis',
+    'data_loading',
     'mapping'
 ]
