--- crested_myna/data_analysis/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class DataAnalysisConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'data_analysis'
