+++ crested_myna/data_loading/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DataAnalysisConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'data_loading'
